+++ app/gui/bp_overview/__init__.py
@@ -1,0 +1,4 @@
+from .bp_overview_tab import BPOverviewTab
+
+
+__all__ = ["BPOverviewTab"]